.config/gdb/python/requirements/large.txt
#!/usr/bin/env -S pip install -r

gpustat
ruff-lsp
torchinfo
